Make diff colour compare against #have instead of #head
When retrieving the 'original' version of the file to generate the diff colours, retrieve the #have version of the file, not the #head revision (which is the default).

This change required changing the type of revision from 'number' to 'string' so that we could pass in 'have'.

NOTE: The 'fragment' parameter was previously not used anywhere